checking to improve performance

+++ kcore/block-level-dual-compact/src/device_funcs.cu
@@ -4,7 +4,7 @@
 #include "./buffer.cc"
 #include "./scans.cc"
 
-__device__ uint64_t blk;
+__device__ unsigned long long int blk;
 
 __device__ void syncBlocks(volatile unsigned int* blockCounter){
